refactor: 리뷰 연결 페이지 이름 수정(ReviewZone)
