+++ frontend/src/pages/ReviewZonePage/styles.ts
@@ -1,13 +1,13 @@
 import styled from '@emotion/styled';
 
-export const ReviewDashboardPage = styled.div`
+export const ReviewZonePage = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
 `;
 
-export const DashboardMainImg = styled.img`
+export const ReviewZoneMainImg = styled.img`
   width: 43rem;
   height: 23rem;
 `;
