+++ frontend/src/pages/index.tsx
@@ -5,4 +5,4 @@ export { default as LoadingPage } from './LoadingPage';
 export { default as ReviewListPage } from './ReviewListPage';
 export { default as ReviewWritingCardFormPage } from './ReviewWritingCardFromPage';
 export { default as ReviewWritingCompletePage } from './ReviewWritingCompletePage';
-export { default as ReviewDashboardPage } from './ReviewDashboardPage';
+export { default as ReviewZonePage } from './ReviewZonePage';
